Move benchmarking `fs` utils to common ones
