+++ src/benchmark/framework/experiment/lightweightItems/lightweightDeserializer.ts
@@ -1,6 +1,12 @@
 import { ModelParams } from "../../../../llm/llmServices/modelParams";
 
 import { deserializeCodeElementRange } from "../../../../utils/codeElementPositions";
+import {
+    getDatasetDir,
+    joinPaths,
+    listJsonFiles,
+    readFile,
+} from "../../../../utils/fs";
 import { throwError } from "../../../../utils/throwErrors";
 import { BenchmarkingLogger } from "../../logging/benchmarkingLogger";
 import { readRequestedFilesCache } from "../../parseDataset/cacheHandlers/cacheReader";
@@ -22,12 +28,6 @@ import {
 } from "../../utils/collectionUtils/mapUtils";
 import { createParamsResolvers } from "../../utils/commonStructuresUtils/llmServicesUtils";
 import { deserializeGoal } from "../../utils/coqUtils/goalParser";
-import {
-    getDatasetDir,
-    joinPaths,
-    listJsonFiles,
-    readFile,
-} from "../../utils/fileUtils/fs";
 
 import { LightweightSerialization } from "./lightweightSerialization";
