+++ src/benchmark/framework/experiment/lightweightItems/lightweightSerializer.ts
@@ -1,6 +1,15 @@
 import { ModelParams } from "../../../../llm/llmServices/modelParams";
 
 import { serializeCodeElementRange } from "../../../../utils/codeElementPositions";
+import {
+    clearDirectory,
+    createDirectory,
+    getDatasetDir,
+    getLastName,
+    joinPaths,
+    relativizeAbsolutePaths,
+    writeToFile,
+} from "../../../../utils/fs";
 import { toFormattedJsonString } from "../../../../utils/printers";
 import { throwError } from "../../../../utils/throwErrors";
 import { BenchmarkingLogger } from "../../logging/benchmarkingLogger";
@@ -23,15 +32,6 @@ import {
 } from "../../utils/collectionUtils/mapUtils";
 import { serializeGoal } from "../../utils/coqUtils/goalParser";
 import { buildSafeJsonFileName } from "../../utils/fileUtils/fileNameUtils";
-import {
-    clearDirectory,
-    createDirectory,
-    getDatasetDir,
-    getLastName,
-    joinPaths,
-    relativizeAbsolutePaths,
-    writeToFile,
-} from "../../utils/fileUtils/fs";
 import { prependWithZeros } from "../../utils/serializationUtils";
 
 import { LightweightSerialization } from "./lightweightSerialization";
