+++ src/benchmark/framework/parseDataset/core/filterTargetsMissingInCache.ts
@@ -1,3 +1,4 @@
+import { listCoqSourceFiles } from "../../../../utils/fs";
 import { stringifyAnyValue } from "../../../../utils/printers";
 import { illegalState } from "../../../../utils/throwErrors";
 import { BenchmarkingLogger } from "../../logging/benchmarkingLogger";
@@ -12,7 +13,6 @@ import { WorkspaceRoot } from "../../structures/common/workspaceRoot";
 import { DatasetCacheUsageMode } from "../../structures/inputParameters/datasetCaching";
 import { ExperimentRunOptions } from "../../structures/inputParameters/experimentRunOptions";
 import { all } from "../../utils/collectionUtils/listUtils";
-import { listCoqSourceFiles } from "../../utils/fileUtils/fs";
 import { readRequestedFilesCache } from "../cacheHandlers/cacheReader";
 import { WorkspaceCacheHolder } from "../cacheStructures/cacheHolders";
 import { createEmptyCache } from "../utils/cacheHoldersUtils";
